OS/2: Make remote command terminate on EOF when reading from stdin.
This didn't work because we use socketpair in OS/2 instead of pipes and in socketpair (no USE_PIPES) mode OpenSSH uses one socket both for stdin and stdout which breaks receiving EOF when stdin is closed on the other end because it remains open as stdout.

Also, this commit adds usual EINTR handling to read/write/waitpid calls added in 1958f184.

Closes #7.